+++ Source/Blazorise/Base/BaseBarToggler.cs
@@ -12,6 +12,8 @@ namespace Blazorise.Base
     {
         #region Members
 
+        private bool isOpen;
+
         #endregion
 
         #region Methods
@@ -19,24 +21,50 @@ namespace Blazorise.Base
         protected override void RegisterClasses()
         {
             ClassMapper
-                .Add( () => ClassProvider.BarToggler() );
+                .Add( () => ClassProvider.BarToggler() )
+                .Add( () => ClassProvider.BarTogglerCollapsed( IsOpen ) );
 
             base.RegisterClasses();
         }
 
         protected void ClickHandler()
         {
-            Clicked?.Invoke();
+            // NOTE: is this right?
+            if ( Clicked == null )
+                ParentBar?.Toggle();
+            else
+                Clicked?.Invoke();
+        }
+
+        protected override void OnInit()
+        {
+            ParentBar?.Hook( this );
+
+            base.OnInit();
         }
 
         #endregion
 
         #region Properties
 
+        [Parameter]
+        internal protected bool IsOpen
+        {
+            get => isOpen;
+            set
+            {
+                isOpen = value;
+
+                ClassMapper.Dirty();
+            }
+        }
+
         /// <summary>
         /// Occurs when the button is clicked.
         /// </summary>
         [Parameter] protected Action Clicked { get; set; }
+
+        [CascadingParameter] protected BaseBar ParentBar { get; set; }
 
         [Parameter] protected RenderFragment ChildContent { get; set; }